Crate solana_storage_proto

👎Deprecated since 3.1.0: This crate has been marked for formal inclusion in the Agave Unstable API. From v4.0.0 onward, the agave-unstable-api crate feature must be specified to acknowledge use of an interface that may break without warning.

Modules

convertDeprecated

Structs

StoredExtendedRewardDeprecated

StoredTokenAmountDeprecated

StoredTransactionStatusMetaDeprecated

StoredTransactionTokenBalanceDeprecated

Type Aliases

StoredExtendedRewardsDeprecated