solana_storage_proto::convert

Module generated

👎Deprecated since 3.1.0: This crate has been marked for formal inclusion in the Agave Unstable API. From v4.0.0 onward, the agave-unstable-api crate feature must be specified to acknowledge use of an interface that may break without warning.

Structs

BlockHeightDeprecated

CompiledInstructionDeprecated

ConfirmedBlockDeprecated

ConfirmedTransactionDeprecated

InnerInstructionDeprecated

InnerInstructionsDeprecated

MessageDeprecated

MessageAddressTableLookupDeprecated

MessageHeaderDeprecated

NumPartitionsDeprecated

ReturnDataDeprecated

RewardDeprecated

RewardsDeprecated

TokenBalanceDeprecated

TransactionDeprecated

TransactionErrorDeprecated

TransactionStatusMetaDeprecated

UiTokenAmountDeprecated

UnixTimestampDeprecated

Enums

RewardTypeDeprecated