solana_storage_proto::convert

Module tx_by_addr

👎Deprecated since 3.1.0: This crate has been marked for formal inclusion in the Agave Unstable API. From v4.0.0 onward, the agave-unstable-api crate feature must be specified to acknowledge use of an interface that may break without warning.

Structs

CustomErrorDeprecated

InstructionErrorDeprecated

MemoDeprecated

TransactionByAddrDeprecated

TransactionByAddrInfoDeprecated

TransactionDetailsDeprecated

TransactionErrorDeprecated

UnixTimestampDeprecated

Enums

InstructionErrorTypeDeprecated

TransactionErrorTypeDeprecated