solana_storage_proto

Type Alias StoredExtendedRewards

pub type StoredExtendedRewards = Vec<StoredExtendedReward>;

👎Deprecated since 3.1.0: This crate has been marked for formal inclusion in the Agave Unstable API. From v4.0.0 onward, the agave-unstable-api crate feature must be specified to acknowledge use of an interface that may break without warning.

Aliased Type

struct StoredExtendedRewards { /* private fields */ }