Enum solana_storage_proto::convert::tx_by_addr::InstructionErrorType

#[repr(i32)]
pub enum InstructionErrorType {

    GenericError,
    InvalidArgument,
    InvalidInstructionData,
    InvalidAccountData,
    AccountDataTooSmall,
    InsufficientFunds,
    IncorrectProgramId,
    MissingRequiredSignature,
    AccountAlreadyInitialized,
    UninitializedAccount,
    UnbalancedInstruction,
    ModifiedProgramId,
    ExternalAccountLamportSpend,
    ExternalAccountDataModified,
    ReadonlyLamportChange,
    ReadonlyDataModified,
    DuplicateAccountIndex,
    ExecutableModified,
    RentEpochModified,
    NotEnoughAccountKeys,
    AccountDataSizeChanged,
    AccountNotExecutable,
    AccountBorrowFailed,
    AccountBorrowOutstanding,
    DuplicateAccountOutOfSync,
    Custom,
    InvalidError,
    ExecutableDataModified,
    ExecutableLamportChange,
    ExecutableAccountNotRentExempt,
    UnsupportedProgramId,
    CallDepth,
    MissingAccount,
    ReentrancyNotAllowed,
    MaxSeedLengthExceeded,
    InvalidSeeds,
    InvalidRealloc,
    ComputationalBudgetExceeded,
    PrivilegeEscalation,
    ProgramEnvironmentSetupFailure,
    ProgramFailedToComplete,
    ProgramFailedToCompile,
    Immutable,
    IncorrectAuthority,
    BorshIoError,
    AccountNotRentExempt,
    InvalidAccountOwner,
    ArithmeticOverflow,
    UnsupportedSysvar,
    IllegalOwner,
    MaxAccountsDataAllocationsExceeded,
    MaxAccountsExceeded,
    MaxInstructionTraceLengthExceeded,
    BuiltinProgramsMustConsumeComputeUnits,

}

Variants

GenericError

InvalidArgument

InvalidInstructionData

InvalidAccountData

AccountDataTooSmall

InsufficientFunds

IncorrectProgramId

MissingRequiredSignature

AccountAlreadyInitialized

UninitializedAccount

UnbalancedInstruction

ModifiedProgramId

ExternalAccountLamportSpend

ExternalAccountDataModified

ReadonlyLamportChange

ReadonlyDataModified

DuplicateAccountIndex

ExecutableModified

RentEpochModified

NotEnoughAccountKeys

AccountDataSizeChanged

AccountNotExecutable

AccountBorrowFailed

AccountBorrowOutstanding

DuplicateAccountOutOfSync

Custom

InvalidError

ExecutableDataModified

ExecutableLamportChange

ExecutableAccountNotRentExempt

UnsupportedProgramId

CallDepth

MissingAccount

ReentrancyNotAllowed

MaxSeedLengthExceeded

InvalidSeeds

InvalidRealloc

ComputationalBudgetExceeded

PrivilegeEscalation

ProgramEnvironmentSetupFailure

ProgramFailedToComplete

ProgramFailedToCompile

Immutable

IncorrectAuthority

BorshIoError

AccountNotRentExempt

InvalidAccountOwner

ArithmeticOverflow

UnsupportedSysvar

IllegalOwner

MaxAccountsDataAllocationsExceeded

MaxAccountsExceeded

MaxInstructionTraceLengthExceeded

BuiltinProgramsMustConsumeComputeUnits

Implementations

impl InstructionErrorType

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of InstructionErrorType.

pub fn from_i32(value: i32) -> Option<InstructionErrorType>

Converts an i32 to a InstructionErrorType, or None if value is not a valid variant.

impl InstructionErrorType

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.